Improve how language is selected
modified:   src/views/__init__.py
modified:   src/views/answer.py
modified:   src/views/previous.py
modified:   src/views/question.py
