+++ src/views/__init__.py
@@ -32,6 +32,12 @@ def validate_language(langs):
     if isinstance(langs, str):
         langs = [langs]
     for lang in langs:
+        # force strings like en-us to correct format
+        try:
+            lang = lang[:2].lower()
+        except:
+            print(f"Got weird lang param: {lang}")
+            return default_language
         if lang in supported_languages:
             return lang
         return default_language
@@ -50,3 +56,19 @@ def select_question_by_language(questionJson, language):
     else:
         print('language not defined')
         return question[default_language]
+
+
+def fix_question_language(question, language):
+    print(question)
+    print(language)
+    if question['type'] == 'multi':
+        question['attribute_question'] = select_question_by_language(
+            question['attribute_question'], language)
+        for attribute in question['attributes']:
+            attribute['attribute_name'] = select_question_by_language(
+                attribute['attribute_name'], language)
+    else:
+        question['attribute_name'] = select_question_by_language(
+            question['attribute_name'], language)
+        question['attribute_question'] = select_question_by_language(
+            question['attribute_question'], language)
